plugins/kernel/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "txl_kernel"
description = "TXL plugin for a kernel driver"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "David Brochart", email = "david.brochart@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "txl",
  "python-dateutil >=2.8.2",
]
dynamic = ["version"]

[project.urls]
Source = "https://github.com/davidbrochart/jpterm/plugins/kernel"

[tool.hatch.version]
path = "txl_kernel/__init__.py"
